remove old_trending
