--- app/modules/trending/old_schemas.py
@@ -1,25 +1,0 @@
-from typing import List, Optional
-from pydantic import BaseModel
-
-
-class TrendingStockKr(BaseModel):
-    num: int
-    ticker: str
-    name: str
-    volume: float
-    current_price: Optional[float]
-    current_price_rate: Optional[float]
-
-
-class TrendingStockUs(BaseModel):
-    num: int
-    ticker: str
-    name: str
-    volume: float
-    current_price: Optional[float]
-    current_price_rate: Optional[float]
-
-
-class TrendingStock(BaseModel):
-    kr: List[TrendingStockKr]
-    us: List[TrendingStockUs]
